Fix a binary-incompatible API change in Interceptor.Chain (#9360)
* Fix a binary-incompatible API change in Interceptor.Chain

* Explicitly depend on Okio

I saw build flakes in CI on JDK8 without this

    file:///home/runner/work/okhttp/okhttp/mockwebserver-deprecated/src/main/kotlin/okhttp3/mockwebserver/DeprecationBridge.kt:156:21
    Symbol is declared in an unnamed module which is not read by current module.

+++ okhttp/src/commonJvmAndroid/kotlin/okhttp3/Interceptor.kt
@@ -107,7 +107,7 @@ fun interface Interceptor {
      * Returns a new chain with the specified connect timeout.
      */
     fun withConnectTimeout(
-      timeout: Long,
+      timeout: Int,
       unit: TimeUnit,
     ): Chain
 
@@ -120,7 +120,7 @@ fun interface Interceptor {
      * Returns a new chain with the specified read timeout.
      */
     fun withReadTimeout(
-      timeout: Long,
+      timeout: Int,
       unit: TimeUnit,
     ): Chain
 
@@ -133,7 +133,7 @@ fun interface Interceptor {
      * Returns a new chain with the specified write timeout.
      */
     fun withWriteTimeout(
-      timeout: Long,
+      timeout: Int,
       unit: TimeUnit,
     ): Chain
