+++ okhttp/src/commonJvmAndroid/kotlin/okhttp3/internal/http/RealInterceptorChain.kt
@@ -172,34 +172,34 @@ class RealInterceptorChain(
   override fun connectTimeoutMillis(): Int = connectTimeoutMillis
 
   override fun withConnectTimeout(
-    timeout: Long,
+    timeout: Int,
     unit: TimeUnit,
   ): Interceptor.Chain {
     check(exchange == null) { "Timeouts can't be adjusted in a network interceptor" }
 
-    return copy(connectTimeoutMillis = checkDuration("connectTimeout", timeout, unit))
+    return copy(connectTimeoutMillis = checkDuration("connectTimeout", timeout.toLong(), unit))
   }
 
   override fun readTimeoutMillis(): Int = readTimeoutMillis
 
   override fun withReadTimeout(
-    timeout: Long,
+    timeout: Int,
     unit: TimeUnit,
   ): Interceptor.Chain {
     check(exchange == null) { "Timeouts can't be adjusted in a network interceptor" }
 
-    return copy(readTimeoutMillis = checkDuration("readTimeout", timeout, unit))
+    return copy(readTimeoutMillis = checkDuration("readTimeout", timeout.toLong(), unit))
   }
 
   override fun writeTimeoutMillis(): Int = writeTimeoutMillis
 
   override fun withWriteTimeout(
-    timeout: Long,
+    timeout: Int,
     unit: TimeUnit,
   ): Interceptor.Chain {
     check(exchange == null) { "Timeouts can't be adjusted in a network interceptor" }
 
-    return copy(writeTimeoutMillis = checkDuration("writeTimeout", timeout, unit))
+    return copy(writeTimeoutMillis = checkDuration("writeTimeout", timeout.toLong(), unit))
   }
 
   override fun withDns(dns: Dns): Interceptor.Chain {
